requirements.txt
opencv-python==4.4.0.46
pandas==1.1.4
pillow==8.0.1
torch==1.7.0
torchvision==0.8.1
tensorboardX==2.1
tqdm==4.50.2